requirements.txt
hyperliquid-python-sdk>=0.10.1
python-dotenv>=1.0.0
eth-account>=0.5.9
requests>=2.28.0
